Make interfaces for EventListToolbar and EventListView plus a little some refactor.
