+++ eventdiscovery-sdk/src/main/java/com/schedjoules/eventdiscovery/eventlist/view/EventListToolbar.java
@@ -17,70 +17,23 @@
 
 package com.schedjoules.eventdiscovery.eventlist.view;
 
-import android.content.res.Resources;
-import android.support.v4.app.FragmentActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
-
-import com.schedjoules.eventdiscovery.R;
-import com.schedjoules.eventdiscovery.common.BaseActivity;
-
-
 /**
- * The toolbar of the Event list screen.
+ * Represents the Toolbar of the Event list screen.
  *
  * @author Gabor Keszthelyi
  */
-public final class EventListToolbar
+public interface EventListToolbar
 {
-    private final Toolbar mToolbar;
-    private final Listener mListener;
+    /**
+     * Sets the title for the toolbar.
+     */
+    void setToolbarTitle(CharSequence title);
 
-
-    public EventListToolbar(Toolbar toolbar, Listener listener)
+    interface Listener
     {
-        mToolbar = toolbar;
-        mListener = listener;
-    }
-
-
-    public void initToolbar(FragmentActivity fragmentActivity)
-    {
-        BaseActivity activity = (BaseActivity) fragmentActivity;
-
-        mToolbar.setTitle(""); // Need to set it to something here, otherwise it will use title from Activity label
-        mToolbar.setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View v)
-            {
-                mListener.onToolbarTitleClick();
-            }
-        });
-        activity.setSupportActionBar(mToolbar);
-
-        Resources res = activity.getResources();
-        if (res.getBoolean(R.bool.schedjoules_enableBackArrowOnEventListScreen))
-        {
-            //noinspection ConstantConditions
-            activity.getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        }
-        else
-        {
-            mToolbar.setTitleMarginStart(res.getDimensionPixelSize(R.dimen.schedjoules_list_item_horizontal_margin));
-        }
-    }
-
-
-    public void setToolbarTitle(CharSequence title)
-    {
-        mToolbar.setTitle(title);
-    }
-
-
-    public interface Listener
-    {
+        /**
+         * Called when the user taps on the toolbar title.
+         */
         void onToolbarTitleClick();
     }
-
 }
